minirefactor of rest tests + init exec rest test
