+++ Portal/vip-api/src/test/java/fr/insalyon/creatis/vip/api/data/ClassesTestUtils.java
@@ -29,11 +29,22 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-B license and that you accept its terms.
  */
-package fr.insalyon.creatis.vip.api.rest.itest.config;
+package fr.insalyon.creatis.vip.api.data;
+
+import fr.insalyon.creatis.vip.application.client.bean.AppClass;
+
+import java.util.ArrayList;
 
 /**
- * Created by abonnet on 7/21/16.
+ * Created by abonnet on 7/26/16.
  */
-public class RestTestUtils {
-    public static final String JSON_CONTENT_TYPE_UTF8 ="application/json;charset=UTF-8";
+public class ClassesTestUtils {
+
+    static public AppClass class1;
+    static public AppClass class2;
+
+    static {
+        class1 = new AppClass("classe 1", new ArrayList<>());
+        class2 = new AppClass("classe 2", new ArrayList<>());
+    }
 }
